+++ lib/src/providers/file_provider.dart
@@ -18,7 +18,8 @@ class FileLottie extends LottieProvider {
     var cacheKey = 'file-${file.path}';
     return sharedLottieCache.putIfAbsent(cacheKey, () async {
       var bytes = await file.readAsBytes();
-      var composition = await LottieComposition.fromBytes(bytes);
+      var composition = await LottieComposition.fromBytes(bytes,
+          name: p.basenameWithoutExtension(file.path));
 
       for (var image in composition.images.values) {
         image.loadedImage ??= await _loadImage(composition, image);