+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.3.0] - 2020-03-02
 - Add `LottieDelegates` a group of options to customize the lottie animation at runtime.
   ie: Dynamically modify color, position, size, text... of every elements of the animation.
+- Correctly display Linear and Radial Gradients
 - Integrate latest changes from Lottie-android
 
 ## [0.2.2] - 2020-02-21